+++ packages/paste-token-designer/constants/Routes.js
@@ -1,0 +1,5 @@
+export const Routes = {
+  EDIT: '/edit',
+  EDIT_BUCKET: bucket => `/edit/${bucket}`,
+  PLAYGROUND: '/playground',
+};
