+++ packages/paste-token-designer/pages/edit/index.js
@@ -1,0 +1,6 @@
+import Link from 'next/link';
+import {EditorWrapper} from '../../components/layout/editor-wrapper';
+
+export default function Home() {
+  return <EditorWrapper>Select an category from the dropdown above to begin editing</EditorWrapper>;
+}
